LayoutTests/platform/mac/accessibility/native-vs-nonnative-checkboxes.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="image" role="checkbox" aria-checked="true" id="check1">check1<BR>
<input type="image" role="checkbox" aria-checked="mixed" id="check2">check2<BR>
<input type="image" role="checkbox" id="check3">check2<BR>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that input types masquerading as checkboxes will use the aria-checked values for its value.");

    if (window.accessibilityController) {

          // aria-state=true
          document.getElementById("check1").focus();
          shouldBe("accessibilityController.focusedElement.intValue", "1");

          // aria-state=false
          document.getElementById("check3").focus();
          shouldBe("accessibilityController.focusedElement.intValue", "0");

          // aria-state=mixed
          document.getElementById("check2").focus();
          shouldBe("accessibilityController.focusedElement.intValue", "2");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
