safari-600.7-branch/LayoutTests/platform/mac/accessibility/native-vs-nonnative-checkboxes-expected.txt

check1
check2
check2
This tests that input types masquerading as checkboxes will use the aria-checked values for its value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.intValue is 1
PASS accessibilityController.focusedElement.intValue is 0
PASS accessibilityController.focusedElement.intValue is 2
PASS successfullyParsed is true

TEST COMPLETE
